requirements.txt
cachetools==5.3.0
certifi==2022.12.7
google-api-python-client==2.86.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
pyasn1==0.5.0
pyparsing==3.0.9
requests==2.29.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15

pip~=23.1.2
cryptography~=3.4.8
lxml~=4.8.0
Pillow~=9.0.1
keyring~=23.5.0
chardet~=4.0.0
setuptools~=67.7.2
future~=0.18.2
python-dotenv~=1.0.0
click~=8.0.3